Moving smoke test files to the root directory
