+++ spec/command_line_helper.rb
