+++ spec/git_helper.rb
