--- tsuru-testing/spec/integration/tsuru-api/tsuru_api_spec.rb
@@ -1,6 +1,0 @@
-require 'serverspec_helper'
-
-describe command('ls /tmp') do
-  its(:exit_status) { should eq 0 }
-end
-
