--- tsuru-testing/spec/integration/tsuru-gandalf/tsuru_gandalf_spec.rb
@@ -1,8 +1,0 @@
-require 'serverspec_helper'
-
-describe file('/home/git/.bash_profile') do
-  it "should use a DNS to connect to the api" do
-    should contain(/export TSURU_HOST=https:\/\/.*-api/)
-  end
-end
-
